+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### Bugs
 * Fix #6038: Support for Gradle configuration cache
+* Fix #6059: Swallow rejected execution from internal usage of the informer executor
 
 #### Improvements
 * Fix #6008: removing the optional dependency on bouncy castle