+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix #6038: Support for Gradle configuration cache
 * Fix #6059: Swallow rejected execution from internal usage of the informer executor
 * Fix #6068: KubernetesMockServer provides incomplete Configuration while creating test Config for KubernetesClient
+* Fix #6085: model getters have same annotations as fields (breaks native)
 
 #### Improvements
 * Fix #6008: removing the optional dependency on bouncy castle